fix(Header): fixing context selector in header ui and code structure and renaming
