+++ client/packages/portal-core/src/context-selector/ContextProvider.tsx
@@ -1,25 +1,14 @@
-
 import { ContextProvider as FusionContextProvider } from '@equinor/fusion-react-context-selector';
 import { ReactNode } from 'react';
-import {
-    useContextResolver,
-} from '../hooks';
-
+import { useContextResolver } from '../hooks';
 
 interface PortalContextProviderProps {
-    children: ReactNode
+	children: ReactNode;
 }
 
 export const ContextProvider = ({ children }: PortalContextProviderProps) => {
-    //Todo: get contest types from view / work-surface
-    const resolver = useContextResolver(['ProjectMaster']);
+	//Todo: get contest types from view / work-surface
+	const resolver = useContextResolver(['ProjectMaster']);
 
-    return (
-        <FusionContextProvider
-            resolver={resolver}
-        >
-            {children}
-        </FusionContextProvider >
-    );
+	return <FusionContextProvider resolver={resolver}>{children}</FusionContextProvider>;
 };
-
