+++ client/packages/portal-core/src/context-selector/index.ts
@@ -1,4 +1,3 @@
-export * from "./ContextSelector"
-export * from "./ContextProvider"
-export * from "./PortalContextSelector"
-export * from "./TopBarContextSelector"
+export * from './ContextSelector';
+export * from './ContextProvider';
+export * from './PortalContextSelector';
