+++ client/packages/portal-ui/src/index.ts
@@ -6,18 +6,18 @@ export * from './lib/full-page-loading';
 export * from './lib/fullscreen-icon/FullscreenIcon';
 export * from './lib/help/Help';
 export * from './lib/my-account/MyAccount';
-export * from './lib/widgets'
-export * from './lib/portal-header/Icon';
-export * from './lib/portal-header/PortalHeader';
+export * from './lib/widgets';
+export * from './lib/header/Icon';
+export * from './lib/header/Header';
 export * from './lib/portal-menu';
 export * from './lib/portal-side-sheet';
 export * from './lib/style-provider/styleProvider';
 export * from './lib/svg-icon-from-string';
 export * from './lib/task/Task';
 export * from './lib/top-bar-actions';
-export * from './lib/breadcrumbs/';
+export * from './lib/header/breadcrumbs';
 export * from './lib/progress-loader';
-export * from "./lib/portal-message-page";
-export * from "./lib/task/work-assigned/WorkAssigned"
-export * from "./lib/side-sheet-header/SideSheetHeader"
-export * from "./lib/kpis";
+export * from './lib/portal-message-page';
+export * from './lib/task/work-assigned/WorkAssigned';
+export * from './lib/side-sheet-header/SideSheetHeader';
+export * from './lib/kpis';
