--- client/packages/portal-ui/src/lib/breadcrumbs/index.ts
@@ -1,1 +1,0 @@
-export * from "./PortalBreadcrumbs"
