+++ client/packages/portal-ui/src/lib/header/HeaderContextSelector.tsx
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+import { ContextSelector } from '@equinor/portal-core';
+export { ContextProvider } from '@equinor/portal-core';
+
+const StyledWrapper = styled.span`
+	width: 45%;
+`;
+
+export const TopBarContextSelector = () => {
+	return (
+		<StyledWrapper>
+			<ContextSelector variant={'header'} />
+		</StyledWrapper>
+	);
+};
