+++ client/packages/portal-ui/src/lib/header/HeaderStyles.ts
@@ -2,15 +2,19 @@ import { TopBar } from '@equinor/eds-core-react';
 import styled from 'styled-components';
 
 export const StyledHeader = styled(TopBar.Header)`
-  grid-template-columns: auto auto auto auto auto auto;
-  gap: 8px;
+	width: max-content;
 `;
 
 export const StyledTopBar = styled(TopBar)`
-  padding: 0px;
-  height: 48px;
+	padding: 0px;
+	height: 48px;
+`;
+
+export const StyledCustomContent = styled(TopBar.CustomContent)`
+	display: flex;
+	align-items: center;
 `;
 
 export const StyledActionsWrapper = styled(TopBar.Actions)`
-  padding-right: 0.5rem;
+	padding-right: 0.5rem;
 `;
