+++ client/packages/portal-ui/src/lib/header/breadcrumbs/index.ts
@@ -1,0 +1,1 @@
+export * from './Breadcrumbs';
